backend/requirements.txt
flask==3.1.0
flask-socketio==5.5.1
flask-cors==5.0.1
eventlet==0.39.1
PyJWT==2.10.1
bcrypt==4.3.0
pandas==2.2.3
requests==2.32.3
selenium==4.27.1
webdriver-manager==4.0.2
openpyxl==3.1.5
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
cryptography==44.0.1
python-dotenv==1.0.1
